# Replace Sortition cpp with Go