@barnjamin 2021-06-22
Feel free to just close this, mostly curious if it'd work and how it'd change perf.