@barnjamin 2021-07-11
needs more testing and might need to work with logprob instead